+++ test/Repository/Pdo/PdoServiceFactoryTest.php
@@ -1,18 +1,18 @@
 <?php
+
 /**
- * @see       https://github.com/zendframework/zend-expressive-authentication-oauth2 for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive-authentication-oauth2/blob/master/LICENSE.md
- *     New BSD License
+ * @see       https://github.com/mezzio/mezzio-authentication-oauth2 for the canonical source repository
+ * @copyright https://github.com/mezzio/mezzio-authentication-oauth2/blob/master/COPYRIGHT.md
+ * @license   https://github.com/mezzio/mezzio-authentication-oauth2/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Expressive\Authentication\OAuth2\Repository\Pdo;
+namespace MezzioTest\Authentication\OAuth2\Repository\Pdo;
 
+use Mezzio\Authentication\OAuth2\Exception;
+use Mezzio\Authentication\OAuth2\Repository\Pdo\PdoService;
+use Mezzio\Authentication\OAuth2\Repository\Pdo\PdoServiceFactory;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
-use Zend\Expressive\Authentication\OAuth2\Exception;
-use Zend\Expressive\Authentication\OAuth2\Repository\Pdo\PdoService;
-use Zend\Expressive\Authentication\OAuth2\Repository\Pdo\PdoServiceFactory;
 
 class PdoServiceFactoryTest extends TestCase
 {