+++ test/ResourceServerFactoryTest.php
@@ -1,19 +1,19 @@
 <?php
+
 /**
- * @see       https://github.com/zendframework/zend-expressive-authentication-oauth2 for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive-authentication-oauth2/blob/master/LICENSE.md
- *     New BSD License
+ * @see       https://github.com/mezzio/mezzio-authentication-oauth2 for the canonical source repository
+ * @copyright https://github.com/mezzio/mezzio-authentication-oauth2/blob/master/COPYRIGHT.md
+ * @license   https://github.com/mezzio/mezzio-authentication-oauth2/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Expressive\Authentication\OAuth2;
+namespace MezzioTest\Authentication\OAuth2;
 
 use League\OAuth2\Server\Repositories\AccessTokenRepositoryInterface;
 use League\OAuth2\Server\ResourceServer;
+use Mezzio\Authentication\OAuth2\Exception;
+use Mezzio\Authentication\OAuth2\ResourceServerFactory;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
-use Zend\Expressive\Authentication\OAuth2\Exception;
-use Zend\Expressive\Authentication\OAuth2\ResourceServerFactory;
 
 class ResourceServerFactoryTest extends TestCase
 {
@@ -47,7 +47,7 @@ class ResourceServerFactoryTest extends TestCase
     }
 
     /**
-     * @expectedException Zend\Expressive\Authentication\OAuth2\Exception\InvalidConfigException
+     * @expectedException Mezzio\Authentication\OAuth2\Exception\InvalidConfigException
      */
     public function testInvokeWithConfigWithoutRepository()
     {